Initial refactoring before fixing evaluation of methods on primitives.
