+++ src/mono/wasm/debugger/BrowserDebugProxy/MonoSDBHelper.cs
@@ -1415,7 +1415,7 @@ namespace Microsoft.WebAssembly.Diagnostics
                 expr = "$\"" + dispAttrStr + "\"";
                 JObject retValue = await resolver.Resolve(expr, token);
                 if (retValue == null)
-                    retValue = await EvaluateExpression.CompileAndRunTheExpression(expr, resolver, token);
+                    retValue = await ExpressionEvaluator.CompileAndRunTheExpression(expr, resolver, token);
 
                 return retValue?["value"]?.Value<string>();
             }
